Update lib tests
- split tests in files per action
- extract common classes for specific files
- add new context for when have 204 status code
